fedora/ks.cfg: Enable gnome autologin
Some things might not be initialized if noone has logged in in the
graphical prompt. This includes pactl sinks not being initialized.

This change will help with automatic tests, the known example being
AUD tests which will sometimes fail to detect devices using pactl
if no user is logged in the graphical login prompt

Signed-off-by: Filip Gołaś <filip.golas@3mdeb.com>